update source.list check

+++ nginx-build.sh
@@ -14,7 +14,6 @@ if [ ! -x /usr/bin/curl ]; then
     apt-get install curl >>/tmp/nginx-ee.log 2>&1
 fi
 
-
 ##################################
 # Variables
 ##################################
@@ -23,11 +22,10 @@ NAXSI_VER=0.56
 OPENSSL_VER=OpenSSL_1_1_1-pre9
 DIR_SRC=/usr/local/src
 NGINX_STABLE=1.14.0
-NGINX_MAINLINE=$(curl -sL https://nginx.org/en/download.html 2>&1 | grep -E -o "nginx\\-[0-9.]+\\.tar[.a-z]*" | awk -F "nginx-" '/.tar.gz$/ {print $2}' | sed -e 's|.tar.gz||g' | head -n 1 2>&1)
-
+NGINX_MAINLINE=$(curl -sL https://nginx.org/en/download.html 2>&1 | grep -E -o 'nginx\-[0-9.]+\.tar[.a-z]*' | awk -F "nginx-" '/.tar.gz$/ {print $2}' | sed -e 's|.tar.gz||g' | head -n 1 2>&1)
 
 # Colors
-CSI="\\033["
+CSI='\033['
 CEND="${CSI}0m"
 CRED="${CSI}1;31m"
 CGREEN="${CSI}1;32m"
@@ -43,13 +41,31 @@ echo "" >/tmp/nginx-ee.log
 while [[ $# -gt 0 ]]; do
     arg="$1"
     case $arg in
-        -p|--pagespeed)  PAGESPEED="y"; shift;;
-        -n|--naxsi)   NAXSI="y"; shift;;
-        -r|--rtmp)   RTMP="y"; shift;;
-        -m|--mainline)   NGINX_RELEASE=1; shift;;
-        -s|--stable)   NGINX_RELEASE=2; shift;;
-        -u|--update) NGINX_RELEASE=2; shift;;
-        *)
+    -p | --pagespeed)
+        PAGESPEED="y"
+        shift
+        ;;
+    -n | --naxsi)
+        NAXSI="y"
+        shift
+        ;;
+    -r | --rtmp)
+        RTMP="y"
+        shift
+        ;;
+    -m | --mainline)
+        NGINX_RELEASE=1
+        shift
+        ;;
+    -s | --stable)
+        NGINX_RELEASE=2
+        shift
+        ;;
+    -u | --update)
+        NGINX_RELEASE=2
+        shift
+        ;;
+    *) ;;
     esac
     shift
 done
@@ -61,28 +77,28 @@ done
 echo ""
 echo "Welcome to the nginx-ee bash script."
 echo ""
-if [ -z  $NGINX_RELEASE ]; then
+if [ -z $NGINX_RELEASE ]; then
     echo ""
     echo "Do you want to compile the latest Nginx Mainline [1] or Stable [2] Release ?"
     while [[ $NGINX_RELEASE != "1" && $NGINX_RELEASE != "2" ]]; do
         read -p "Select an option [1-2]: " NGINX_RELEASE
     done
 fi
-if [ -z  $PAGESPEED ]; then
+if [ -z $PAGESPEED ]; then
     echo ""
     echo "Do you want Ngx_Pagespeed ? (y/n)"
     while [[ $PAGESPEED != "y" && $PAGESPEED != "n" ]]; do
         read -p "Select an option [y/n]: " PAGESPEED
     done
 fi
-if [ -z  $NAXSI ]; then
+if [ -z $NAXSI ]; then
     echo ""
     echo "Do you want NAXSI WAF (still experimental)? (y/n)"
     while [[ $NAXSI != "y" && $NAXSI != "n" ]]; do
         read -p "Select an option [y/n]: " NAXSI
     done
 fi
-if [ -z  $RTMP ]; then
+if [ -z $RTMP ]; then
     echo ""
     echo "Do you want RTMP streaming module ?"
     while [[ $RTMP != "y" && $RTMP != "n" ]]; do
@@ -127,15 +143,15 @@ fi
 # Install dependencies
 ##################################
 
-echo -ne "       Installing dependencies               [..]\\r"
+echo -ne '       Installing dependencies               [..]\r'
 apt-get update >>/tmp/nginx-ee.log 2>&1
 apt-get install -y git build-essential libtool automake autoconf zlib1g-dev \
-libpcre3-dev libgd-dev libssl-dev libxslt1-dev libxml2-dev libgeoip-dev \
-libgoogle-perftools-dev libperl-dev libpam0g-dev libxslt1-dev libbsd-dev zip unzip gnupg gnupg2 >>/tmp/nginx-ee.log 2>&1
+    libpcre3-dev libgd-dev libssl-dev libxslt1-dev libxml2-dev libgeoip-dev \
+    libgoogle-perftools-dev libperl-dev libpam0g-dev libxslt1-dev libbsd-dev zip unzip gnupg gnupg2 >>/tmp/nginx-ee.log 2>&1
 
 if [ $? -eq 0 ]; then
     echo -ne "       Installing dependencies                [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "        Installing dependencies              [${CRED}FAIL${CEND}]"
     echo ""
@@ -158,8 +174,6 @@ if [ ! -d /etc/nginx ]; then
 
 fi
 
-
-
 ##################################
 # Install gcc7 or gcc8 from PPA
 ##################################
@@ -174,10 +188,10 @@ fi
 # install gcc-7
 distro_version=$(lsb_release -sc)
 
-if [[ "$NGINX_RELEASE" = "1" && "$RTMP" = "n" ]]; then
-    if [[ "$distro_version" == "xenial" || "$distro_version" == "bionic" ]]; then
-        if [[ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-bionic.list || ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-xenial.list ]]; then
-            echo -ne "       Installing gcc-8                       [..]\\r"
+if [ "$NGINX_RELEASE" = "1" ] && [ "$RTMP" = "n" ]; then
+    if [ "$distro_version" == "xenial" ] || [ "$distro_version" == "bionic" ]; then
+        if [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-bionic.list ] && [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-8_1-xenial.list ]; then
+            echo -ne '       Installing gcc-8                       [..]\r'
             {
                 apt-get install software-properties-common -y
                 add-apt-repository ppa:jonathonf/gcc-8.1 -y
@@ -189,7 +203,7 @@ if [[ "$NGINX_RELEASE" = "1" && "$RTMP" = "n" ]]; then
             export CXX="/usr/bin/gc++-8"
             if [ $? -eq 0 ]; then
                 echo -ne "       Installing gcc-8                       [${CGREEN}OK${CEND}]\\r"
-                echo -ne "\\n"
+                echo -ne '\n'
             else
                 echo -e "        Installing gcc-8                      [${CRED}FAIL${CEND}]"
                 echo ""
@@ -202,7 +216,7 @@ if [[ "$NGINX_RELEASE" = "1" && "$RTMP" = "n" ]]; then
 else
     if [ "$distro_version" == "xenial" ]; then
         if [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-gcc-7_1-xenial.list ]; then
-            echo -ne "       Installing gcc-7                       [..]\\r"
+            echo -ne '       Installing gcc-7                       [..]\r'
             {
                 apt-get install software-properties-common -y
                 add-apt-repository ppa:jonathonf/gcc-7.1 -y
@@ -214,7 +228,7 @@ else
             export CXX="/usr/bin/gc++-7"
             if [ $? -eq 0 ]; then
                 echo -ne "       Installing gcc-7                       [${CGREEN}OK${CEND}]\\r"
-                echo -ne "\\n"
+                echo -ne '\n'
             else
                 echo -e "        Installing gcc-7                      [${CRED}FAIL${CEND}]"
                 echo ""
@@ -236,7 +250,7 @@ fi
 ##################################
 
 if [ "$RTMP" = "y" ]; then
-    echo -ne "       Installing FFMPEG for RMTP module      [..]\\r"
+    echo -ne '       Installing FFMPEG for RMTP module      [..]\r'
     {
         if [ "$distro_version" == "xenial" ]; then
             if [ ! -f /etc/apt/sources.list.d/jonathonf-ubuntu-ffmpeg-4-xenial.list ]; then
@@ -251,7 +265,7 @@ if [ "$RTMP" = "y" ]; then
     } >>/tmp/nginx-ee.log 2>&1
     if [ $? -eq 0 ]; then
         echo -ne "       Installing FFMPEG for RMTP module      [${CGREEN}OK${CEND}]\\r"
-        echo -ne "\\n"
+        echo -ne '\n'
     else
         echo -e "       Installing FFMPEG for RMTP module      [${CRED}FAIL${CEND}]"
         echo ""
@@ -270,7 +284,7 @@ fi
 cd $DIR_SRC || exit
 rm -rf $DIR_SRC/*.tar.gz $DIR_SRC/nginx-1.* ipscrubtmp ipscrub
 
-echo -ne "       Downloading additionals modules        [..]\\r"
+echo -ne '       Downloading additionals modules        [..]\r'
 
 {
     # cache_purge module
@@ -362,7 +376,7 @@ echo -ne "       Downloading additionals modules        [..]\\r"
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading additionals modules        [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "        Downloading additionals modules      [${CRED}FAIL${CEND}]"
     echo ""
@@ -377,7 +391,7 @@ fi
 
 cd $DIR_SRC || exit
 
-echo -ne "       Downloading brotli                     [..]\\r"
+echo -ne '       Downloading brotli                     [..]\r'
 {
     if [ -d $DIR_SRC/ngx_brotli ]; then
         { git -C $DIR_SRC/ngx_brotli pull origin master; }
@@ -390,7 +404,7 @@ echo -ne "       Downloading brotli                     [..]\\r"
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading brotli                     [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "       Downloading brotli      [${CRED}FAIL${CEND}]"
     echo ""
@@ -403,7 +417,7 @@ fi
 # Download OpenSSL
 ##################################
 
-echo -ne "       Downloading openssl                    [..]\\r"
+echo -ne '       Downloading openssl                    [..]\r'
 
 cd $DIR_SRC || exit
 {
@@ -420,7 +434,7 @@ cd $DIR_SRC || exit
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading openssl                    [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "       Downloading openssl      [${CRED}FAIL${CEND}]"
     echo ""
@@ -435,7 +449,7 @@ fi
 
 cd $DIR_SRC || exit
 if [ "$NAXSI" = "y" ]; then
-    echo -ne "       Downloading naxsi                      [..]\\r"
+    echo -ne '       Downloading naxsi                      [..]\r'
     {
         if [ -d $DIR_SRC/naxsi ]; then
             rm -rf $DIR_SRC/naxsi
@@ -447,7 +461,7 @@ if [ "$NAXSI" = "y" ]; then
 
     if [ $? -eq 0 ]; then
         echo -ne "       Downloading naxsi                      [${CGREEN}OK${CEND}]\\r"
-        echo -ne "\\n"
+        echo -ne '\n'
     else
         echo -e "       Downloading naxsi      [${CRED}FAIL${CEND}]"
         echo ""
@@ -464,7 +478,7 @@ fi
 
 cd $DIR_SRC || exit
 if [ "$PAGESPEED" = "y" ]; then
-    echo -ne "       Downloading pagespeed                  [..]\\r"
+    echo -ne '       Downloading pagespeed                  [..]\r'
 
     {
         rm -rf incubator-pagespeed-ngx-latest-beta build_ngx_pagespeed.sh install
@@ -475,7 +489,7 @@ if [ "$PAGESPEED" = "y" ]; then
 
     if [ $? -eq 0 ]; then
         echo -ne "       Downloading pagespeed                  [${CGREEN}OK${CEND}]\\r"
-        echo -ne "\\n"
+        echo -ne '\n'
     else
         echo -e "       Downloading pagespeed                  [${CRED}FAIL${CEND}]"
         echo ""
@@ -490,7 +504,7 @@ fi
 ##################################
 
 cd $DIR_SRC || exit
-echo -ne "       Downloading nginx                      [..]\\r"
+echo -ne '       Downloading nginx                      [..]\r'
 if [ -d $DIR_SRC/nginx ]; then
     rm -rf $DIR_SRC/nginx
 fi
@@ -504,7 +518,7 @@ cd $DIR_SRC/nginx/ || exit
 
 if [ $? -eq 0 ]; then
     echo -ne "       Downloading nginx                      [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "       Downloading nginx      [${CRED}FAIL${CEND}]"
     echo ""
@@ -517,7 +531,7 @@ fi
 # Apply Nginx patches
 ##################################
 
-echo -ne "       Applying nginx patches                 [..]\\r"
+echo -ne '       Applying nginx patches                 [..]\r'
 
 wget -O nginx__dynamic_tls_records.patch https://raw.githubusercontent.com/cujanovic/nginx-dynamic-tls-records-patch/master/nginx__dynamic_tls_records_1.13.0%2B.patch >>/tmp/nginx-ee.log 2>&1
 patch -p1 <nginx__dynamic_tls_records.patch >>/tmp/nginx-ee.log 2>&1
@@ -526,7 +540,7 @@ patch -p1 <nginx__dynamic_tls_records.patch >>/tmp/nginx-ee.log 2>&1
 
 if [ $? -eq 0 ]; then
     echo -ne "       Applying nginx patches                 [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "       Applying nginx patches                 [${CRED}FAIL${CEND}]"
     echo ""
@@ -537,62 +551,62 @@ fi
 
 ##################################
 # Configure Nginx
-##################################
+##################################"${nginx_cc_opt[@]}" \
 
-echo -ne "       Configuring nginx                      [..]\\r"
+echo -ne '       Configuring nginx                      [..]\r'
 
 ./configure \
-$ngx_naxsi \
+    $ngx_naxsi
 "${nginx_cc_opt[@]}" \
 --with-ld-opt='-Wl,-Bsymbolic-functions -fPIE -pie -Wl,-z,relro -Wl,-z,now' \
---prefix=/usr/share/nginx \
---conf-path=/etc/nginx/nginx.conf \
---http-log-path=/var/log/nginx/access.log \
---error-log-path=/var/log/nginx/error.log \
---lock-path=/var/lock/nginx.lock \
---pid-path=/var/run/nginx.pid \
---http-client-body-temp-path=/var/lib/nginx/body \
---http-fastcgi-temp-path=/var/lib/nginx/fastcgi \
---http-proxy-temp-path=/var/lib/nginx/proxy \
---http-scgi-temp-path=/var/lib/nginx/scgi \
---http-uwsgi-temp-path=/var/lib/nginx/uwsgi \
---with-pcre-jit \
---with-http_ssl_module \
---with-http_stub_status_module \
---with-http_realip_module \
---with-http_auth_request_module \
---with-http_addition_module \
---with-http_geoip_module \
---with-http_gzip_static_module \
---with-http_image_filter_module \
---with-http_v2_module \
---with-http_sub_module \
---with-http_xslt_module \
---with-file-aio \
---with-threads \
---add-module=/usr/local/src/ngx_cache_purge \
---add-module=/usr/local/src/memc-nginx-module \
---add-module=/usr/local/src/ngx_devel_kit \
---add-module=/usr/local/src/headers-more-nginx-module \
---add-module=/usr/local/src/echo-nginx-module \
---add-module=/usr/local/src/ngx_http_substitutions_filter_module \
---add-module=/usr/local/src/redis2-nginx-module \
---add-module=/usr/local/src/srcache-nginx-module \
---add-module=/usr/local/src/set-misc-nginx-module \
---add-module=/usr/local/src/ngx_http_redis \
---add-module=/usr/local/src/ngx_brotli \
---add-module=/usr/local/src/ipscrub \
---add-module=/usr/local/src/ngx_http_auth_pam_module \
---add-module=/usr/local/src/nginx-module-vts \
-$ngx_pagespeed \
-$ngx_rtmp \
---with-openssl=/usr/local/src/openssl \
---with-openssl-opt=enable-tls1_3 \
---sbin-path=/usr/sbin/nginx >>/tmp/nginx-ee.log 2>&1
+    --prefix=/usr/share/nginx \
+    --conf-path=/etc/nginx/nginx.conf \
+    --http-log-path=/var/log/nginx/access.log \
+    --error-log-path=/var/log/nginx/error.log \
+    --lock-path=/var/lock/nginx.lock \
+    --pid-path=/var/run/nginx.pid \
+    --http-client-body-temp-path=/var/lib/nginx/body \
+    --http-fastcgi-temp-path=/var/lib/nginx/fastcgi \
+    --http-proxy-temp-path=/var/lib/nginx/proxy \
+    --http-scgi-temp-path=/var/lib/nginx/scgi \
+    --http-uwsgi-temp-path=/var/lib/nginx/uwsgi \
+    --with-pcre-jit \
+    --with-http_ssl_module \
+    --with-http_stub_status_module \
+    --with-http_realip_module \
+    --with-http_auth_request_module \
+    --with-http_addition_module \
+    --with-http_geoip_module \
+    --with-http_gzip_static_module \
+    --with-http_image_filter_module \
+    --with-http_v2_module \
+    --with-http_sub_module \
+    --with-http_xslt_module \
+    --with-file-aio \
+    --with-threads \
+    --add-module=/usr/local/src/ngx_cache_purge \
+    --add-module=/usr/local/src/memc-nginx-module \
+    --add-module=/usr/local/src/ngx_devel_kit \
+    --add-module=/usr/local/src/headers-more-nginx-module \
+    --add-module=/usr/local/src/echo-nginx-module \
+    --add-module=/usr/local/src/ngx_http_substitutions_filter_module \
+    --add-module=/usr/local/src/redis2-nginx-module \
+    --add-module=/usr/local/src/srcache-nginx-module \
+    --add-module=/usr/local/src/set-misc-nginx-module \
+    --add-module=/usr/local/src/ngx_http_redis \
+    --add-module=/usr/local/src/ngx_brotli \
+    --add-module=/usr/local/src/ipscrub \
+    --add-module=/usr/local/src/ngx_http_auth_pam_module \
+    --add-module=/usr/local/src/nginx-module-vts \
+    $ngx_pagespeed \
+    $ngx_rtmp \
+    --with-openssl=/usr/local/src/openssl \
+    --with-openssl-opt=enable-tls1_3 \
+    --sbin-path=/usr/sbin/nginx >>/tmp/nginx-ee.log 2>&1
 
 if [ $? -eq 0 ]; then
     echo -ne "       Configuring nginx                      [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "        Configuring nginx    [${CRED}FAIL${CEND}]"
     echo ""
@@ -605,7 +619,7 @@ fi
 # Compile Nginx
 ##################################
 
-echo -ne "       Compiling nginx                        [..]\\r"
+echo -ne '       Compiling nginx                        [..]\r'
 
 {
     make -j "$(nproc)"
@@ -614,7 +628,7 @@ echo -ne "       Compiling nginx                        [..]\\r"
 
 if [ $? -eq 0 ]; then
     echo -ne "       Compiling nginx                        [${CGREEN}OK${CEND}]\\r"
-    echo -ne "\\n"
+    echo -ne '\n'
 else
     echo -e "        Compile nginx      [${CRED}FAIL${CEND}]"
     echo ""
